remove simulatenous evaluations
The idea behind this was to evaluate multiple different datasets at once. However this did not work as the cluster always tried to find nodes matching all requirements. = it tried to find nodes matching all evaluations. It tried to run all pods on the same node. This did ofc not work. Therefor, this option is removed by me